tests/wpt/tests/navigation-api/state/location-reload.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
async_test(t => {
  window.onload = t.step_func(() => {
    i.contentWindow.navigation.navigate("#", { history: "replace", state: { data: "value" } });
    assert_equals(i.contentWindow.navigation.entries().length, 1);
    assert_equals(i.contentWindow.navigation.currentEntry.getState().data, "value");

    i.contentWindow.location.reload();
    i.onload = t.step_func_done(() => {
      assert_equals(i.contentWindow.navigation.entries().length, 1);
      assert_equals(i.contentWindow.navigation.currentEntry.getState().data, "value");
    });
  });
}, "entry.getState() after location.reload()");
</script>
